+++ assets/javascripts/kitten/components/structure/cards/team-card/components/button-with-tooltip.js
@@ -68,7 +68,7 @@ export const TeamCardButtonWithTooltip = ({
 
     <Button
       fit="icon"
-      size="tiny"
+      size="small"
       modifier="hydrogen"
       tag="span"
       aria-label={phoneNumber}