+++ assets/javascripts/kitten/components/structure/expand/expand-board/index.js
@@ -5,6 +5,7 @@ import { Button } from '../../../action/button'
 import { ArrowIcon } from '../../../graphics/icons/arrow-icon'
 import COLORS from '../../../../constants/colors-config'
 import { pxToRem } from '../../../../helpers/utils/typography'
+import { checkDeprecatedSizes } from '../../../../helpers/utils/deprecated'
 
 const StyledButton = styled(Button)`
   width: 100%;
@@ -46,7 +47,7 @@ export class ExpandBoardButton extends PureComponent {
     onClick: () => {},
     ariaId: '',
     borderRadius: 0,
-    size: 'big',
+    size: 'large',
   }
 
   render() {
@@ -61,6 +62,8 @@ export class ExpandBoardButton extends PureComponent {
       size,
     } = this.props
 
+    checkDeprecatedSizes(size)
+    
     const defaultExpandChildren = expandChildren ? expandChildren : children
 
     return (