+++ assets/javascripts/kitten/components/typography/sup-title/index.js
@@ -3,7 +3,7 @@ import { Text } from '../../typography/text'
 
 export const SupTitle = props => (
   <Text
-    size="tiny"
+    size="small"
     lineHeight="normal"
     transform="uppercase"
     weight="bold"