+++ CHANGELOG.md
@@ -7,7 +7,17 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Breaking change: `Label`: 
   - Replace `small` prop size default value.
   - Remove `nano` prop size value.
- 
+- Change/Deprecation: 
+  - Replace `tiny, regular, default, big` by `small, medium, large` 
+    prop size values to the following folders:
+  - `Action`,
+  - `Typography`,
+  - `Form`,
+  - `Graphics`,
+  - `Information`,
+  - `Structure`,
+  - `Layer`
+
 ## [9.6.0] - 2022-04-04
 
 Features: